% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_in_stability.R
\name{change_in_stability}
\alias{change_in_stability}
\title{change in stability}
\usage{
change_in_stability(rna_pol2_data, percent_cutoff = 15, plot_label = "sample")
}
\arguments{
\item{rna_pol2_data}{tibble, with 5 columns, 1. gene_id, 2. pol2_control 3.
pol2_test 4. rna_control 5. rna_test}

\item{percent_cutoff}{numeric, cutoff to determine actively transcribing
genes, i.e. to consider top 10 or top 15 percent genes. Default: 15
@param plot_label Title for plots. For eg. "test Vs Control" or "ros Vs unstressed"}
}
\value{
a tibble of all genes denoted as being stabilized or destabilized in
test condition
}
\description{
Change in stability determined by applying linear regression on
RNAP II and mRNA levels in two conditions.
}
\details{
Linear regression and studentized residual method is used to classify the genes based on their
fold-change in mRNA and RNAP II levels. Only actively transcribing genes
(by RNAP II) in either test or control are considered.
}
\examples{
\dontrun{
if(interactive()){
 rna_pol2_data <- readr::read_delim(pipe("pbpaste"), delim="\t", col_names = TRUE)
 }
}
}
