% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_from_htseqcount.R
\name{deseq_from_htseqcount}
\alias{deseq_from_htseqcount}
\title{deseq_from_htseqcount}
\usage{
deseq_from_htseqcount(
  dir,
  pattern,
  metadata_file,
  HTSeqOutput = TRUE,
  drop_genes = NULL,
  minimum_read_cutoff = 2,
  outfile
)
}
\arguments{
\item{dir}{string, absolute path of directory containing read count file for
each sample (a HTSeqCount output)}

\item{pattern}{string, pattern of files to be selected as input from the dir.
For example `pattern="*._HTSeqCount.txt`
  The pattern is also used to replace the pattern name from file, which is
  used as column names in further processing. The pattern replaced column
  names should match with first column of metadata file.}

\item{metadata_file}{string, absolute path of metadata file in tab-delimited
  format. metadata file should contain two columns, First: names of the
  samples exactly matching with the pattern replaced file names. Second:
  sample type i.e. treated or untreated
sample1_set1    treated}

\item{HTSeqOutput}{logical, whether the read count files are HTSeq-count
output or not. If TRUE, last six rows from containing the statistics are
ignored. If FALSE, the file is considered as it is. Default: TRUE}

\item{drop_genes}{string, pattern of the genes to be dropped (Like
mitochondrial rRNA, rRNA, tRNAs) from calculation.For example:'CaglfM.*'
mitochrondrial rRNA in C. glabrata. Default:NULL}

\item{minimum_read_cutoff}{numeric, remove genes containing reads less than
the threshold in all samples from further processing, Default: 2}

\item{outfile}{string, name of the output file}
}
\value{
combined count matrix of all samples, log2 correlation plot and
  expression matrix
}
\description{
Computes normalized gene expression value and differential gene
  expression using DESeq2 from read counts matrix obtained from HTSeqCount or
  any other method.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 dir = (".")
 pattern="_12.*_HTSeqCount.txt"
 metadata = system.file("exdata", "metadata.txt", package = "FungalTranscriptStability")
 deseq_from_htseqcount(dir = dir, pattern = pattern, metadata_file = metadata_file, drop_genes = "CaglfM*", outfile = "ROS")
 }
}

}
