% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_codon_distribution.R
\name{optimal_codon_distribution}
\alias{optimal_codon_distribution}
\title{optimal_codon_distribution}
\usage{
optimal_codon_distribution(
  optimal_codon_table,
  genelist_stabilityInfo,
  ref_sequence_file
)
}
\arguments{
\item{optimal_codon_table}{Codon table, containing pre-defined optimal and
Non-optimal codons}

\item{genelist_stabilityInfo}{tibble, of genes with stability information.
First_column: gene_id, second_column: stability ie. whether the gene is
stable or unstable, third_column: condition ie. stressed or unstressed}

\item{ref_sequence_file}{path or connection of coding nucleotides sequences
fasta file. The number of sequences should be equal to number of genes in
genelist_stabilityInfo tibble. Sequences can be subsetted from all genes
sequences using fastaR::fa_some_records() function.}
}
\value{
a tibble of optimal codon percentage for each gene, boxplot of codon
distribution in different class or stability of genes.
}
\description{
Codon distribution based on pre-defined optimal codons for yeast
species (Pechmann, 2013)
}
\examples{
\dontrun{
if(interactive()){
 codon_table <- system.file("extdata", "codonOptimalityTable_Yeasts.txt", package = "FungalTranscriptStability")
 org_codon_table <- readr::read_tsv(codon_table) \%>\% dplyr::select(Codon, Cgla)
 ref_sequence_file <- system.file("extdata", "CgROS_stability.fa", package = "FungalTranscriptStability")
 }
}
}
