% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orf_len.R
\name{orf_len_distribution}
\alias{orf_len_distribution}
\title{orf_len_distribution}
\usage{
orf_len_distribution(gff, genelist_stabilityInfo)
}
\arguments{
\item{gff}{path or connection of gff file}

\item{genelist_stabilityInfo}{tibble, of genes with stability information.
First_column: gene_id, second_column: stability ie. whether the gene is
stable or unstable, third_column: condition ie. stressed or unstressed}
}
\value{
pvalue of the orf lengths, boxplot of orf length distribution
}
\description{
displays distribution of orf lengths for give stable and
unstable genes.
}
\examples{
\dontrun{
if(interactive()){
  gff_file <- "/Users/Pooja/Documents/PhD/thesis_figures/Chapter3/C_glabrata_CBS138_version_s02-m07-r38_features.gff"
 }
}
}
