% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_go_in_diamond.R
\name{plot_go_in_diamond}
\alias{plot_go_in_diamond}
\title{plot_go_in_diamond}
\usage{
plot_go_in_diamond(data_tbl)
}
\arguments{
\item{data_tbl}{tibble, a tibble of GO-terms and additional columns as 1.
GO-term, 2. percent of genes enriched over background, 3. pvalue, 4.type
ie. stable or unstable, 5. class ie. ros, infected etc.}
}
\value{
GO term enrichment plot
}
\description{
allows to visualize GO term enrichment across different
categories and samples
}
\examples{
\dontrun{
if(interactive()){
 data <- readr::read_delim(pipe("pbpaste"), delim="\t", col_names =T)
 }
}
}
