% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(
  dat_mat,
  color_key = c(-1, 0, 1),
  top_annotation = TRUE,
  key_name = "expression_level",
  show_row_names = FALSE,
  additional_column = NULL,
  additional_column_key = c(-8, 0, 8)
)
}
\arguments{
\item{dat_mat}{matrix of input data to be plotted, all the columns should be numeric.}

\item{color_key}{numeric vector, determine range for color key. Continuous colors for non-negative value. And discriminate color for negative values.
Default: c(-1, 0, 1)}

\item{top_annotation}{logical, to plot top annotation i.e. boxplot or not. Default: TRUE}

\item{additional_column}{matrix, to plot additional heatmap. Default: NULL}

\item{additional_column_key}{numeric vector, determine range for color key for additional heatmap. Default: c(-8, 0, 8)}
}
\value{
heatmap object
}
\description{
plot heatmap
}
\examples{
\dontrun{
if(interactive()){
 dat_mat1 <- readr::read_delim(pipe("pbpaste"), delim="\t", col_names = TRUE)
 dd <- dat_mat1 \%>\% dplyr::select(-c("rank_ratio"))  \%>\% tibble::column_to_rownames(var="gene_id")
 add_col <-  dat_mat1 \%>\% dplyr::select(c("gene_id","rank_ratio")) \%>\% tibble::column_to_rownames(var="gene_id")
 }
}
}
