% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabilityofLowlyTranscribingGenes.R
\name{stabilityofLowlyTranscribingGenes}
\alias{stabilityofLowlyTranscribingGenes}
\title{stabilityofLowlyTranscribingGenes}
\usage{
stabilityofLowlyTranscribingGenes(
  rna_pol2_tbl,
  percent_cutoff = 85,
  exp_cutoff = 2,
  plot_heatmap = FALSE,
  color_key = c(-0.75, 0, 0.75)
)
}
\arguments{
\item{rna_pol2_tbl}{a tibble of normalized RNAP II and mRNA values. With
first column (character) of gene-id, second column (numeric) RNAP II values
and third column (numeric) mRNA values.}

\item{percent_cutoff}{numeric, cutoff to determine actively transcribing
genes, i.e. to consider bottom 85 percent genes. Default: 85}

\item{exp_cutoff}{numeric, minimum expression value in either RNAP II or mRNA
to filter. Default: 2}

\item{plot_heatmap}{logical, to plot heatmap of the z-score of RNAP II and
mRNA values.If FALSE tibble will be returned. Default: FALSE}

\item{color_key}{numeric vector, range for heatmap color key.  Default:
c(-0.75, 0, 0.75)}
}
\value{
quantile analysis plot, heatmap and tibble containing stable and
  unstable genes.
}
\description{
Stability of lowly or non-transcribing genes is determined by
  comparing RNAP II occupancy levels and steady state mRNA levels
}
\examples{
\dontrun{
if(interactive()){
 rna_pol2_tbl <- readr::read_delim(pipe("pbpaste"), delim="\t", col_names = TRUE)
 }
}

}
