% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpm_from_htseqcount.R
\name{tpm_from_htseqcount}
\alias{tpm_from_htseqcount}
\title{tpm_from_htseqcount}
\usage{
tpm_from_htseqcount(
  dir,
  pattern,
  gff_file,
  HTSeqOutput = TRUE,
  drop_genes = NULL,
  minimum_read_cutoff = 2,
  write_output = FALSE,
  output_name
)
}
\arguments{
\item{dir}{string, path of the text files output from htseq-count}

\item{pattern}{string, determining pattern to detect the count file and to
remove extensions from the sample names}

\item{gff_file}{string, path or connection to gff file}

\item{HTSeqOutput}{logical, whether the output is Htseq-count or not. If it's
htseq-count output last sixlines are filtered from further process.
Default: TRUE}

\item{drop_genes}{string, to drop genes with specific pattern like
mitochondrial, Default: NULL}

\item{minimum_read_cutoff}{numeric, filter genes which has less than cut-off
reads in all of the samples , Default: 2}

\item{write_output}{logical, whether to return output on console or write in
a file. If true, output file is generated. Default: FALSE}

\item{output_name}{string, name or path of output file name.}
}
\value{
a correlation plot and tpm matrix of input files.
}
\description{
TPM can be calculated from count matrix. Count matrix can be
  output of htseq-count or prepDE in Hisat2-Stringtie pipeline.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 gff_file <- "/Users/Pooja/Documents/PhD/thesis_figures/Chapter3/C_glabrata_CBS138_version_s02-m07-r38_features.gff"
 pattern="_12.*_HTSeqCount.txt"
 dir="/Users/Pooja/Documents/PhD/thesis_figures/Chapter3/rna"
 tpm_from_htseqcount(dir = dir, pattern = pattern,gff_file=gff_file, HTSeqOutput = TRUE, write_output = TRUE, output_name = "tpm_ros_rna_repeat1",drop_genes = "CaglfM*" )
 }
}
}
